Add sites to readme, add mentors list
Update the readme to include links to our NodeSchool site and Meetup
site, and reduce the information contained within it (since the
information is on those sites).  Also include a mentors list which
folks can add to as the group grows in size.

(Most of this format was copied from the London NodeSchool repo.)